perf(fast_time): use u64 nanosecond clock arithmetic on Linux
Linux clock_gettime_nanos and TimeSourceImpl cache fields used u128, but
u64 nanoseconds spans ~584 years, far beyond CLOCK_MONOTONIC_COARSE range.
Switching to u64 removes a u64::try_from narrowing and replaces u128
multiply/add/compare with u64 equivalents.

Callgrind: clock_now 108 -> 88 instructions (-18.5%). Windows already used
u64 (get_tick_count_64), so no mirrored change is needed.

+++ packages/fast_time/src/pal/linux/bindings/facade.rs
@@ -26,7 +26,7 @@ impl BindingsFacade {
 
 impl Bindings for BindingsFacade {
     #[inline]
-    fn clock_gettime_nanos(&self) -> u128 {
+    fn clock_gettime_nanos(&self) -> u64 {
         match self {
             Self::Real(bindings) => bindings.clock_gettime_nanos(),
             #[cfg(test)]
